pyproject.toml
[project]
name = "the-rest-of-us"
version = "0.1.0"
description = "Automated daily AI tech news podcast — two hosts, sceptical optimism, no cheerleading"
requires-python = ">=3.11"
dependencies = [
    "anthropic",
    "feedparser",
    "google-genai",
    "httpx",
    "beautifulsoup4",
    "pydub",
    "tavily-python",
    "python-dotenv",
]

[project.optional-dependencies]
rhythm = [
    "whisperx",
    "yt-dlp",
    "librosa",
    "numpy",
    "scipy",
]

[project.scripts]
hn-signal = "hn_signal.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hn_signal"]
